Add files via upload
Lab 3 was a dungeon crawler! This lab was a ton of fun. Building on previous concepts, we had a basic X by X dungeon filled with traps, treasure, and an exit. This focused on randomly generating array elements, checking bounds via accessing those elements, validating user input against array positions, and safely passing arrays. Lots of comments on this one as there are quite a few moving parts.